Unit tests and organization
Refactored source code directory
Changed naming conventions for unit tests
Created a CSV file loader
