+++ Planetarium/src/main/java/com/revature/planetarium/Main.java
@@ -11,22 +11,22 @@ import java.util.Scanner;
 public class Main {
 
 	public static void main(String[] args) {
-		javafy();
-//		try {
-//			AppConfig.configureAppProperties(args);
-//			Javalin app = Javalin.create(config ->{
-//				config.bundledPlugins.enableCors(cors -> {
-//					cors.addRule(it -> {
-//						it.anyHost();
-//					});
-//				});
-//				config.bundledPlugins.enableDevLogging();
-//			});
-//			JavalinSetup.mapRoutes(app);
-//			app.start(8080);
-//		} catch (ConfigurationFail e) {
-//			e.printStackTrace();
-//		}
+//		javafy();
+		try {
+			AppConfig.configureAppProperties(args);
+			Javalin app = Javalin.create(config ->{
+				config.bundledPlugins.enableCors(cors -> {
+					cors.addRule(it -> {
+						it.anyHost();
+					});
+				});
+				config.bundledPlugins.enableDevLogging();
+			});
+			JavalinSetup.mapRoutes(app);
+			app.start(8080);
+		} catch (ConfigurationFail e) {
+			e.printStackTrace();
+		}
 	}
 
 	private static void javafy() {
@@ -40,7 +40,7 @@ public class Main {
 			if (line.equals("STOP"))
 				break;
 			builder.append(line);
-			if (count % 3 == 2)
+			if (count % 2 == 1)
 				builder.append("\"}, { \"");
 			else
 				builder.append("\", \"");
