+++ Planetarium/src/test/java/com/revature/integration/repository/moon/negative/DaoCreateMoonTest.java
@@ -1,0 +1,46 @@
+package com.revature.integration.repository.moon.negative;
+
+import com.revature.integration.repository.moon.MoonDaoUtil;
+import org.junit.Before;
+import org.junit.Test;
+
+public class DaoCreateMoonTest extends MoonDaoUtil {
+
+
+
+//TODO: each of these should have sad path tests
+//    Nomenclature should go by root name, plus the specified error
+//    e.g., "createMoonNameAlreadyExists"
+
+
+    @Before
+    public void setUp() throws Exception {
+
+    }
+
+    @Test
+    public void createMoon() {
+    }
+    @Test
+    public void readMoonById() {
+    }
+    @Test
+    public void readMoonByName() {
+    }
+    @Test
+    public void readAllMoons() {
+    }
+    @Test
+    public void readMoonsByPlanet() {
+    }
+    @Test
+    public void updateMoon() {
+    }
+    @Test
+    public void deleteMoonById() {
+    }
+    @Test
+    public void deleteMoonByName() {
+    }
+
+}
