+++ Planetarium/src/test/java/com/revature/integration/repository/user/negative/DaoFindUserTest.java
@@ -1,0 +1,5 @@
+package com.revature.integration.repository.user.negative;
+
+public class DaoFindUserTest {
+
+}
