--- Planetarium/src/test/java/com/revature/unitTests/repository/moon/MoonDaoNegativeTest.java
@@ -1,45 +1,0 @@
-package com.revature.unitTests.repository.moon;
-
-import org.junit.Before;
-import org.junit.Test;
-
-public class MoonDaoNegativeTest extends MoonDaoUtil {
-
-
-
-//TODO: each of these should have sad path tests
-//    Nomenclature should go by root name, plus the specified error
-//    e.g., "createMoonNameAlreadyExists"
-
-
-    @Before
-    public void setUp() throws Exception {
-
-    }
-
-    @Test
-    public void createMoon() {
-    }
-    @Test
-    public void readMoonById() {
-    }
-    @Test
-    public void readMoonByName() {
-    }
-    @Test
-    public void readAllMoons() {
-    }
-    @Test
-    public void readMoonsByPlanet() {
-    }
-    @Test
-    public void updateMoon() {
-    }
-    @Test
-    public void deleteMoonById() {
-    }
-    @Test
-    public void deleteMoonByName() {
-    }
-
-}
