--- Planetarium/src/test/java/com/revature/unitTests/service/MoonServiceTest.java
@@ -1,30 +1,0 @@
-package com.revature.unitTests.service;
-
-import org.junit.Test;
-
-
-public class MoonServiceTest {
-    @Test
-    public void createMoon() {
-    }
-
-    @Test
-    public void selectMoon() {
-    }
-
-    @Test
-    public void selectAllMoons() {
-    }
-
-    @Test
-    public void selectByPlanet() {
-    }
-
-    @Test
-    public void updateMoon() {
-    }
-
-    @Test
-    public void deleteMoon() {
-    }
-}
