--- Planetarium/src/test/java/com/revature/unitTests/service/PlanetServiceTest.java
@@ -1,30 +1,0 @@
-package com.revature.unitTests.service;
-
-import org.junit.Test;
-
-public class PlanetServiceTest {
-
-    @Test
-    public void createPlanet() {
-    }
-
-    @Test
-    public void selectPlanet() {
-    }
-
-    @Test
-    public void selectAllPlanets() {
-    }
-
-    @Test
-    public void selectByOwner() {
-    }
-
-    @Test
-    public void updatePlanet() {
-    }
-
-    @Test
-    public void deletePlanet() {
-    }
-}
