--- Planetarium/src/test/java/com/revature/unitTests/service/UserServiceTest.java
@@ -1,13 +1,0 @@
-package com.revature.unitTests.service;
-
-import org.junit.Test;
-
-public class UserServiceTest {
-    @Test
-    public void createUser() {
-    }
-
-    @Test
-    public void authenticate() {
-    }
-}
