+++ Planetarium/src/test/java/com/revature/unitTests/service/planet/PlanetServiceTest.java
@@ -1,0 +1,30 @@
+package com.revature.unitTests.service.planet;
+
+import org.junit.Test;
+
+public class PlanetServiceTest {
+
+    @Test
+    public void createPlanet() {
+    }
+
+    @Test
+    public void selectPlanet() {
+    }
+
+    @Test
+    public void selectAllPlanets() {
+    }
+
+    @Test
+    public void selectByOwner() {
+    }
+
+    @Test
+    public void updatePlanet() {
+    }
+
+    @Test
+    public void deletePlanet() {
+    }
+}
